+++ packages/Angular/Generic/forms/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @memberjunction/ng-forms
 
+## 5.48.0
+
+### Patch Changes
+
+- @memberjunction/ai-core-plus@5.48.0
+- @memberjunction/ng-markdown@5.48.0
+- @memberjunction/ng-ui-components@5.48.0
+
+## 5.47.0
+
+### Patch Changes
+
+- @memberjunction/ai-core-plus@5.47.0
+- @memberjunction/ng-markdown@5.47.0
+- @memberjunction/ng-ui-components@5.47.0
+
+## 5.46.0
+
+### Patch Changes
+
+- @memberjunction/ai-core-plus@5.46.0
+- @memberjunction/ng-markdown@5.46.0
+- @memberjunction/ng-ui-components@5.46.0
+
+## 5.45.1
+
+### Patch Changes
+
+- Updated dependencies [572d219]
+  - @memberjunction/ai-core-plus@5.45.1
+  - @memberjunction/ng-markdown@5.45.1
+  - @memberjunction/ng-ui-components@5.45.1
+
+## 5.45.0
+
+### Patch Changes
+
+- Updated dependencies [13716e4]
+- Updated dependencies [6125dcd]
+- Updated dependencies [ad9f4a3]
+- Updated dependencies [c1f2d3d]
+  - @memberjunction/ng-ui-components@5.45.0
+  - @memberjunction/ai-core-plus@5.45.0
+  - @memberjunction/ng-markdown@5.45.0
+
 ## 5.44.0
 
 ### Patch Changes